19464 create mobile claims proxy (pre-req for caching in worker) (#6157)
* moves controller logic to a claims proxy so that it can be also used from a worker
